requirements.txt
TikTokApi>=5.2.0
requests>=2.28.0
aiohttp>=3.8.0
playwright>=1.40.0
discord.py>=2.3.0
